backend/requirements.txt
  fastapi>=0.104.0
  pydantic==2.7.1
  uvicorn
  spacy>=3.0.0
  opencv-python>=4.0.0
  pytesseract>=0.3.0
  Pillow>=9.0.0
  tensorflow-cpu>=2.10.0
  scikit-learn>=1.0.0
  sentence-transformers>=2.0.0
  tf-keras
  torch>=2.0.0 # Note: For GPU support, you might need to install a specific CUDA-enabled version of torch. Refer to PyTorch documentation.
  PyMuPDF>=1.22.0
  python-docx>=0.8.0
  Werkzeug>=2.0.0
  matplotlib>=3.0.0
  seaborn>=0.12.0
  pandas>=2.0.0
  numpy>=1.20.0
  flask
  tensorflow_hub
  python-multipart
  bcrypt
  PyJwt
  sentence_transformers